+++ x/graphics/image/image.h
@@ -449,8 +449,8 @@ namespace graphics::image
 
     core::error image_load(std::string path, image_data_t & img);
 
-    image_type image_get_type_from_ext(const char * ext, int32_t length = -1);
-    image_type image_get_type(const byte_t * buffer, int32_t length);
+    image_type image_get_type_from_ext(const char * ext, size_t length = -1);
+    image_type image_get_type(const byte_t * buffer, size_t length);
 
     core::error image_create(const byte_t * buffer, int32_t length, image_data_t & img);
     core::error image_create(image_codec_context & ictx, const byte_t * buffer, int32_t length, image_data_t & img);
